Move explorer links to one file
