+++ packages/frontend/src/components/auction/TransactionSuccess/TransactionSuccess.tsx
@@ -2,7 +2,7 @@ import { Transactions } from '@/blockchain/transaction'
 import styled from 'styled-components'
 import { TxFlowSteps } from '../TxFlowSteps'
 import { TransactionSuccessHeader } from './TransactionSuccessHeader'
-import { useExplorerTxLink } from '@/blockchain/hooks/useExplorerLink'
+import { useExplorerTxLink } from '@/blockchain/hooks/useExplorerLinks'
 import { shortenTxHash } from './shortenTxHash'
 import { CopyButton } from '@/components/buttons/CopyButton'
 import { RedirectButton } from '@/components/buttons/RedirectButton'
